pyproject.toml
[tool.poetry]
name = "fastapi-ddd"
version = "0.1.0"
description = "Fastapi DDD with SQLAlchemy"
authors = ["izagood <ljbfif50@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = {extras = ["all"], version = "^0.111.0"}
gunicorn = "^22.0.0"
uvicorn = {extras = ["standard"], version = "^0.30.1"}
loguru = "^0.7.2"
asyncio = "^3.4.3"
sqlalchemy = "^2.0.31"
psycopg2-binary = "^2.9.9"
python-dotenv = "^1.0.1"
alembic = "^1.13.2"

[tool.poetry.group.dev]
optional = true

[tool.poetry.group.dev.dependencies]
ruff = "^0.4.4"
pytest = "^8.2.0"
pytest-mock = "^3.14.0"
pytest-asyncio = "^0.23.6"
pytest-cov = "^5.0.0"

[tool.ruff]
line-length = 120
indent-width = 4

target-version = "py312"

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-ra -q"
pythonpath = ["."]
testpaths = [
    "tests",
]
filterwarnings = ["ignore::DeprecationWarning", "ignore::PendingDeprecationWarning"]

[tool.coverage.run]
branch = true
source = ["fastapi_ddd"]
command_line = "-m pytest"

[tool.coverage.report]
ignore_errors = true
skip_empty = true

[tool.coverage.html]
directory = "coverage_html_report"

[tool.coverage.json]
output = "coverage.json"

[tool.coverage.xml]
output = "coverage.xml"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
